backend/email-receiver/requirements.txt
fastapi==0.115.13
google-api-python-client==2.170.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
python-dotenv==1.1.0